+++ GridKit/Model/PhasorDynamics/Controller/REECB/Reecb.cpp
@@ -1,0 +1,30 @@
+/**
+ * @file Reecb.cpp
+ * @author Luke Lowery (lukel@tamu.edu)
+ * @brief Non-Enzyme instantiation for the REECB electrical-control model.
+ */
+
+#include "ReecbImpl.hpp"
+
+namespace GridKit
+{
+  namespace PhasorDynamics
+  {
+    namespace Controller
+    {
+      /**
+       * @brief Report that a separate Jacobian is unavailable in the plain-real build.
+       */
+      template <typename scalar_type, typename index_type>
+      int Reecb<scalar_type, index_type>::evaluateJacobian()
+      {
+        Log::misc() << "Evaluate Jacobian for Reecb...\n";
+        Log::misc() << "Jacobian evaluation is not implemented!\n";
+        return 0;
+      }
+
+      template class Reecb<double, long int>;
+      template class Reecb<double, size_t>;
+    } // namespace Controller
+  } // namespace PhasorDynamics
+} // namespace GridKit