+++ tests/IntegrationTests/PhasorDynamics/runPDIntegrationTests.cpp
@@ -12,6 +12,7 @@ int main()
   result += test.twoBusTgov1();
   result += test.threeBusBasic();
   result += test.threeBusClassical();
+  result += test.regcaReecbRepca();
 
   return result.summary();
 }