+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 - Remove unnecessary data copying while evaluating `PowerElectronics` models, speeding up large simulations by up to 3x
 - Added `HYGOV` governor model implementation for PhasorDynamics.
 - Added `REPCA` controller model implementation for PhasorDynamics.
+- Added `REECB` electrical-control model implementation for PhasorDynamics.
 
 ## v0.1